+++ packages/verifier/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @bifold/verifier
 
+## 3.0.20
+
+### Patch Changes
+
+- @bifold/react-hooks@3.0.20
+
+## 3.0.19
+
+### Patch Changes
+
+- @bifold/react-hooks@3.0.19
+
+## 3.0.18
+
+### Patch Changes
+
+- @bifold/react-hooks@3.0.18
+
+## 3.0.17
+
+### Patch Changes
+
+- @bifold/react-hooks@3.0.17
+
 ## 3.0.16
 
 ### Patch Changes